docker_scripts/requirements.txt
itis-dakota
osparc-filecomms
pydantic
pydantic-settings==2.5.2
watchdog
